Move secondary chain spec into primary node, remove extra indirection for CLI data structure
