+++ cumulus/node/src/lib.rs
@@ -1,7 +1,4 @@
 //! Cirrus Node CLI library.
 
-mod chain_spec;
-#[macro_use]
-pub mod service;
-pub mod cli;
 mod rpc;
+pub mod service;
